FRESTYLE-369 refactor: AutoMigrate を明示 DDL へ置き換え go.mod から GORM を落とす（段 C・完結） (#2322)
* refactor: 型付け専用スキーマをアプリ中核の実 DDL へ昇格させ本番の実型に合わせる

sqlc の型付け専用だった queries/schema.sql を infra/database/schema/core.sql へ移し、
そのまま起動時に流せる DDL にする（CREATE TABLE / INDEX を IF NOT EXISTS に、
採番列を bigserial に、これまで AutoMigrate が作っていた索引を明記）。
ナレッジ基盤の 2 ファイルと同じく、宣言（sqlc）と実体（DDL）を 1 ファイルに集約する。

あわせて、本番の information_schema と突き合わせて食い違っていた型を実型に合わせた。

- roles.id / users.role_id: smallint → integer（本番・AutoMigrate とも integer）
- company_applications の 4 列: text → varchar（200 / 120 / 255 / 16）
- users.role_id に DEFAULT 3（ローリングデプロイ中の旧コード向けの安全弁）

roles.id の型変更で sqlc の Go 型が int16 → int32 になるため、呼び出し側も追随させる。


* refactor: 起動時マイグレーションを AutoMigrate から明示 DDL へ差し替え *sql.DB へ移す

Migrate / NewPostgres が *gorm.DB ではなく *sql.DB を扱うようにし、AutoMigrate を
schema/core.sql の go:embed 適用（ApplyCoreSchema）へ置き換える。接続は pgx の
stdlib ドライバで開き、pgbouncer 経由のときは simple query protocol を強制する
（これまで GORM 側で立てていた設定と同じ挙動）。

適用順序・冪等性・直列化は据え置き。

- 順序: 中核スキーマ → seed / バックフィル / 明示制約 → ナレッジ基盤 → 権限 → テナント橋渡し
- 冪等: CREATE TABLE / INDEX は IF NOT EXISTS、seed は ON CONFLICT DO NOTHING
- 直列化: これまでと同じキーの pg_advisory_xact_lock で各段を囲む（共通の withMigrateTx へ集約）
- RESET_DB=true の DROP SCHEMA public CASCADE → CREATE SCHEMA public も据え置き

testsupport は GORM のハンドルを被せる足場（gormOn / closeGormPool / AutoMigrate ラッパ）を
撤去し、起動時と同じ明示 DDL でテスト DB を組む。結合テストは薄いラッパを経由せず
database.* を ctx 付きで直接呼ぶ。


* refactor: domain から GORM タグと TableName を撤去しドキュメントを追随させる

スキーマの正本が infra/database/schema/*.sql へ移ったので、domain の gorm タグと
TableName() は何も参照しない死んだメタデータになる。撤去する（json タグはそのまま。
json:"-" による機密の隠蔽も変えていない）。

タグを前提にしていた説明（AutoMigrate が列を足す / GORM を通さない 等）と、
README・archlint のメッセージも実態に合わせる。docs/ は swaggo の再生成分。


* chore: go.mod から gorm を落とす

参照が 1 つも残っていないため go mod tidy で gorm.io/gorm と gorm.io/driver/postgres が
直接依存から外れる。DB ドライバは pgx/v5 の stdlib のみになる。


* fix: ロール ID を列の実型 integer に揃え 16bit への縮小をやめる

roles.id / users.role_id を本番の実型 integer に直した結果、sqlc が int32 を
生成するようになり、domain 側の uint16 へ代入する箇所が 32bit → 16bit の
縮小変換になっていた。ロールは 1〜3 の固定採番なので実害は出ていないが、
エラーも出さずに値が化ける型の穴なので塞ぐ。

domain の Role.ID / RoleID 定数 / User.RoleID を int32 にし、変換自体を無くす。
不要になった int32(...) 6 箇所も除去（unconvert が検出）。


* docs: backend README のデータアクセス方針を要約しテスト戦略の記述を落とす

README はアプリケーションの説明に限定し、テスト戦略は書かない方針に合わせる。


* fix: 起動時スキーマ適用が中核テーブルのロックを握らないようにする

CREATE INDEX IF NOT EXISTS は既存索引でスキップする場合でもテーブルの ShareLock を取り、
トランザクション終了まで手放さない。全 DDL を 1 トランザクションで流していたため、索引を
持つ中核テーブルすべての ShareLock を同時に握った状態で起動時マイグレーションが走り、
その間 INSERT / UPDATE / DELETE が詰まっていた。ALTER TABLE ADD COLUMN IF NOT EXISTS も
同様で、列が既に在っても AccessExclusiveLock を取る（読み取りまで止まる）。

- 索引はカタログ（pg_indexes）を先に引き、未作成のものだけ CREATE INDEX を発行する
- テナント橋渡しの workspace_id 列も、列の存在を見てから ALTER する
- session_notes の一意索引も同じく事前チェックしてから作る
- withMigrateTx に lock_timeout を入れ、ロック待ちを有限にする
  （テーブル 3s / 先行タスク待ち 30s。超過時は指数バックオフで 4 回やり直してから
   起動を失敗させる。無限に待って全ライターを詰まらせないため）

実測（本番相当＝索引が揃った DB、notes へ書き込みトランザクションを 1 本開いたまま）:
修正前は Migrate が notes の ShareLock 待ちで進まず、45 秒の打ち切りまでに
roles / user_oidc_identities / master_exercise_examples の ShareLock を握ったまま、
並行 INSERT 176 件が lock timeout で落ちた。修正後は 35ms で完走し、pg_locks に
衝突するロックは 1 度も現れず、並行 INSERT 116 件はすべて成功した。


* test: DDL の文分割と索引名の抽出に単体テストを足す

事前チェックの前提になる 2 つの純粋関数を固定する。文分割がコメントや引用符の内側の
セミコロンで切れると起動時に壊れた SQL を流すことになり、索引名の抽出を取りこぼすと
既存索引にも CREATE INDEX を発行して ShareLock を握る。埋め込み済みの正本 DDL
3 ファイルが実際に分割でき、CREATE INDEX を認識できることも合わせて確かめる。


* test: 起動時マイグレーション Migrate 自体の結合テストを足す

testsupport.OpenTestDB は Migrate を呼ばず Apply*Schema / Seed* を自前の順序で並べ直して
いるため、Migrate の適用順を壊しても結合テストは緑のまま通っていた（実起動だけが
relation "workspaces" does not exist で落ちる）。

まっさらな public schema へ Migrate を 2 回続けて流し、適用順・冪等性・seed の結果と、
索引が揃った状態の再適用が書き込みを止めないことを固定する。

適用順を入れ替える変異で落ちることを実行確認した:
  - ナレッジ基盤とテナント橋渡しを入れ替え → relation "workspaces" does not exist
  - 中核スキーマを seed の後ろへ移動     → relation "companies" does not exist


* test: 起動 DDL 適用後の列型を結合テストで固定する

本番の実列に合わせて直した型（roles.id / users.role_id / master_exercises.sort_order を
integer、company_applications の 4 列を varchar）を CI に残す。手で 1 回突き合わせた事実が
どこにも残っていないと、元へ戻す変更が誰にも気付かれずに通ってしまう。

CREATE TABLE IF NOT EXISTS は既に在るテーブルの列を作り直さないため、public schema を
作り直してから Migrate を流した姿を見る。

型を戻す変異で落ちることを実行確認した（roles.id → smallint / users.role_id → smallint /
sort_order → bigint / varchar → text / varchar(255) → varchar(320)）。


* ci: sqlc 生成物の drift チェックを足す

schema/*.sql は起動時に流す実 DDL そのものなので、列の型を変えたのに再生成を忘れると
「実スキーマと Go の型が食い違ったまま」マージされる。OpenAPI の drift チェックと同じ形で、
make sqlc 後に生成物へ差分が出ないことを検証する。sqlc の版は Makefile の SQLC_VERSION
固定を使う（版がずれると常に差分が出るため、出どころを 1 つにする）。

roles.id を smallint へ戻す変異で、生成物 4 ファイルに差分が出ること（＝ CI が落ちること）
を実行確認した。


* test: pooler 判定 isPgBouncerDSN に table-driven テストを足す

本番が simple query protocol で喋るかを決めている唯一のスイッチなのにテストが無かった。
false へ倒れると本番でだけ prepared statement does not exist が出て、ローカルと CI では
再現しない。真になる 8 件（pooler ホスト / pgbouncer=true / key=value 形式 / 大文字 /
前後空白）と偽になる 11 件（直結 / パスワード・path・クエリ値・dbname に紛れた文字列 /
壊れた URL / 裸のホスト名）を並べる。

常に false を返す変異で真の 8 件が、常に true を返す変異で偽の 11 件が落ちることを
実行確認した。


* test: nil 判定の後に明示 return を置き静的解析の誤検知を避ける

CI の staticcheck が t.Fatalf の非復帰を追えず、SA5011（nil ポインタ参照の
可能性）を報告して落ちる。ローカルでは同版・同設定で再現しないが、解析器の
環境差に左右されない書き方にしておく。


---------

+++ backend/internal/adapter/persistence/sqlcgen/company_learning_activity.sql.go
@@ -39,7 +39,7 @@ ORDER BY agg.last_active_date DESC NULLS LAST, u.id ASC
 
 type ListCompanyMemberActivitiesParams struct {
 	CompanyID     int64
-	TraineeRoleID int16
+	TraineeRoleID int32
 	FromDate      time.Time
 }
